UNR-CS-425-426-T48-main/UNR-CS-425-426-T48-main/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
coverage==7.4.3
Django==4.2.11
gunicorn==21.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.21
PySocks==1.7.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
djangorestframework==3.15.1
